restlin/requirements.txt
click==7.1.2
fastapi==0.62.0
h11==0.11.0
psycopg2==2.8.6
pydantic==1.7.3
python-engineio==3.13.2
python-socketio==4.6.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.12.1
